+++ activerecord/test/cases/associations/belongs_to_associations_test.rb
@@ -125,6 +125,24 @@ class BelongsToAssociationsTest < ActiveRecord::TestCase
     ActiveRecord::Base.belongs_to_required_by_default = original_value
   end
 
+  def test_required_belongs_to_message_config
+    original_value = ActiveRecord::Base.belongs_to_required_by_default
+    ActiveRecord::Base.belongs_to_required_by_default = true
+
+    model = Class.new(ActiveRecord::Base) do
+      self.table_name = "accounts"
+      def self.name; "Temp"; end
+      belongs_to :company, message: "should be present."
+    end
+
+    account = model.new
+    assert_not_predicate account, :valid?
+    assert_equal [{ error: :blank }], account.errors.details[:company]
+    assert_equal ["should be present."], account.errors[:company]
+  ensure
+    ActiveRecord::Base.belongs_to_required_by_default = original_value
+  end
+
   def test_default
     david = developers(:david)
     jamis = developers(:jamis)